+++ packages/gatsby-plugin-page-creator/src/gatsby-node.ts
@@ -1,5 +1,5 @@
-import glob from "globby"
 import _ from "lodash"
+import glob from "globby"
 import systemPath from "path"
 import { sync as existsSync } from "fs-exists-cached"
 import {
@@ -67,17 +67,19 @@ Please pick a path to an existing directory.`)
     const pagesGlob = `**/*.{${exts}}`
 
     // Get initial list of files.
-    let files = await glob(pagesGlob, { cwd: pagesPath })
+    const files = await glob(pagesGlob, { cwd: pagesPath })
     files.forEach(file => {
       createPage(file, pagesDirectory, actions, ignore, graphql, reporter)
     })
+
+    const knownFiles = new Set(files)
 
     watchDirectory(
       pagesPath,
       pagesGlob,
       addedPath => {
         try {
-          if (!_.includes(files, addedPath)) {
+          if (!knownFiles.has(addedPath)) {
             createPage(
               addedPath,
               pagesDirectory,
@@ -86,7 +88,7 @@ Please pick a path to an existing directory.`)
               graphql,
               reporter
             )
-            files.push(addedPath)
+            knownFiles.add(addedPath)
           }
         } catch (e) {
           reporter.panic(
@@ -108,7 +110,7 @@ Please pick a path to an existing directory.`)
               })
             }
           })
-          files = files.filter(f => f !== removedPath)
+          knownFiles.delete(removedPath)
         } catch (e) {
           reporter.panic(
             e.message.startsWith(`PageCreator`)
@@ -135,7 +137,7 @@ export function setFieldsOnGraphQLNodeType({
 }: SetFieldsOnGraphQLNodeTypeArgs): object {
   try {
     const extensions = store.getState().program.extensions
-    const collectionQuery = `all${type.name}`
+    const collectionQuery = _.camelCase(`all ${type.name}`)
     if (knownCollections.has(collectionQuery)) {
       return {
         gatsbyPath: {