+++ packages/gatsby-recipes/src/renderer/render.js
@@ -51,7 +51,9 @@ const Wrapper = ({
     >
       <ResourceProvider value={plan}>
         <InputProvider value={inputs}>
-          <Suspense fallback={<p>Loading recipe...</p>}>{children}</Suspense>
+          <Suspense fallback={<p>Loading recipe 1/2...</p>}>
+            {children}
+          </Suspense>
         </InputProvider>
       </ResourceProvider>
     </GlobalsProvider>
@@ -347,5 +349,4 @@ const render = (recipe, cb, context = {}, isApply, isStream, name) => {
   }
 }
 
-module.exports.render = render
-module.exports.ResourceComponent = ResourceComponent
+export { render, ResourceComponent }