+++ packages/gatsby-transformer-sharp/src/customize-schema.js
@@ -31,6 +31,7 @@ const {
   PotraceType,
   ImageFitType,
 } = require(`./types`)
+const { prefixId, CODES } = require(`./error-utils`)
 
 function toArray(buf) {
   const arr = new Array(buf.length)
@@ -441,8 +442,13 @@ const createFields = ({
             // this is no longer in progress
             inProgressCopy.delete(publicPath)
             if (err) {
-              console.error(
-                `error copying file from ${details.absolutePath} to ${publicPath}`,
+              reporter.panic(
+                {
+                  id: prefixId(CODES.MissingResource),
+                  context: {
+                    sourceMessage: `error copying file from ${details.absolutePath} to ${publicPath}`,
+                  },
+                },
                 err
               )
             }