+++ packages/gatsby/cache-dir/socketIo.js
@@ -77,7 +77,6 @@ export default function socketIo() {
         // errors within the console, such as when exiting the develop process.
         socket.on(`disconnect`, () => {
           console.warn(`[socket.io] Disconnected from dev server.`)
-          socket.close()
         })
       } catch (err) {
         console.error(`Could not connect to socket.io on dev server.`)