+++ packages/gatsby/src/query/redirects-writer.ts
@@ -13,7 +13,10 @@ export const writeRedirects = async (): Promise<void> => {
   const { program, redirects } = store.getState()
 
   // Filter for redirects that are meant for the browser.
-  const browserRedirects = redirects.filter(r => r.redirectInBrowser)
+  const browserRedirects = redirects
+    .filter(r => r.redirectInBrowser)
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    .map(({ redirectInBrowser, isPermanent, ...rest }) => rest)
 
   const newHash = crypto
     .createHash(`md5`)