+++ packages/babel-preset-gatsby/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.14](https://github.com/gatsbyjs/gatsby/compare/babel-preset-gatsby@0.5.13...babel-preset-gatsby@0.5.14) (2020-10-13)
+
+**Note:** Version bump only for package babel-preset-gatsby
+
+## [0.5.13](https://github.com/gatsbyjs/gatsby/compare/babel-preset-gatsby@0.5.12...babel-preset-gatsby@0.5.13) (2020-10-06)
+
+**Note:** Version bump only for package babel-preset-gatsby
+
+## [0.5.12](https://github.com/gatsbyjs/gatsby/compare/babel-preset-gatsby@0.5.11...babel-preset-gatsby@0.5.12) (2020-10-01)
+
+**Note:** Version bump only for package babel-preset-gatsby
+
+## [0.5.11](https://github.com/gatsbyjs/gatsby/compare/babel-preset-gatsby@0.5.10...babel-preset-gatsby@0.5.11) (2020-09-28)
+
+**Note:** Version bump only for package babel-preset-gatsby
+
 ## [0.5.10](https://github.com/gatsbyjs/gatsby/compare/babel-preset-gatsby@0.5.9...babel-preset-gatsby@0.5.10) (2020-09-15)
 
 **Note:** Version bump only for package babel-preset-gatsby