+++ packages/gatsby-legacy-polyfills/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-legacy-polyfills@0.0.4...gatsby-legacy-polyfills@0.0.5) (2020-09-28)
+
+**Note:** Version bump only for package gatsby-legacy-polyfills
+
 ## [0.0.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-legacy-polyfills@0.0.3...gatsby-legacy-polyfills@0.0.4) (2020-08-28)
 
 **Note:** Version bump only for package gatsby-legacy-polyfills