+++ packages/gatsby-plugin-google-analytics/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.19](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@2.3.18...gatsby-plugin-google-analytics@2.3.19) (2020-10-16)
+
+### Bug Fixes
+
+- **gatsby-plugin-google-analytics:** Fix gatsby-node so we still have a warning on older gatsby versions ([#27495](https://github.com/gatsbyjs/gatsby/issues/27495)) ([9797828](https://github.com/gatsbyjs/gatsby/commit/9797828a5ec686a367937e60ff0400521be1ae68))
+
+## [2.3.18](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@2.3.17...gatsby-plugin-google-analytics@2.3.18) (2020-10-12)
+
+### Bug Fixes
+
+- **gatsby-plugin-google-analytics:** remove required on trackingId ([#27398](https://github.com/gatsbyjs/gatsby/issues/27398)) ([3518d10](https://github.com/gatsbyjs/gatsby/commit/3518d10c3541277a3ba2da5bb099a54d02179257))
+
+## [2.3.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@2.3.16...gatsby-plugin-google-analytics@2.3.17) (2020-10-07)
+
+### Features
+
+- **gatsby:** plugin option validation ([#27242](https://github.com/gatsbyjs/gatsby/issues/27242)) ([9b01ca7](https://github.com/gatsbyjs/gatsby/commit/9b01ca7926efa1c4c6d58b33a137c2f0a0ef99b7))
+
+## [2.3.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@2.3.15...gatsby-plugin-google-analytics@2.3.16) (2020-10-06)
+
+**Note:** Version bump only for package gatsby-plugin-google-analytics
+
+## [2.3.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@2.3.14...gatsby-plugin-google-analytics@2.3.15) (2020-09-28)
+
+**Note:** Version bump only for package gatsby-plugin-google-analytics
+
 ## [2.3.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@2.3.13...gatsby-plugin-google-analytics@2.3.14) (2020-09-15)
 
 **Note:** Version bump only for package gatsby-plugin-google-analytics