+++ packages/gatsby-plugin-preact/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-preact@4.0.15...gatsby-plugin-preact@4.0.16) (2020-10-16)
+
+### Bug Fixes
+
+- **deps:** update minor and patch for gatsby-plugin-preact ([#27392](https://github.com/gatsbyjs/gatsby/issues/27392)) ([1496bf1](https://github.com/gatsbyjs/gatsby/commit/1496bf17971858b16ed18ad742046bca2b8ea3e5))
+
+## [4.0.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-preact@4.0.14...gatsby-plugin-preact@4.0.15) (2020-10-08)
+
+### Bug Fixes
+
+- **gatsby-plugin-preact:** add babel-plugin properly ([#27165](https://github.com/gatsbyjs/gatsby/issues/27165)) ([91a5876](https://github.com/gatsbyjs/gatsby/commit/91a5876f13091dd4c5e690362d03f93885c3bcd7))
+
+## [4.0.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-preact@4.0.13...gatsby-plugin-preact@4.0.14) (2020-10-06)
+
+**Note:** Version bump only for package gatsby-plugin-preact
+
+## [4.0.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-preact@4.0.12...gatsby-plugin-preact@4.0.13) (2020-09-28)
+
+### Bug Fixes
+
+- **gatsby-plugin-preact:** use new babel plugin for better context compat ([#26956](https://github.com/gatsbyjs/gatsby/issues/26956)) ([7f38209](https://github.com/gatsbyjs/gatsby/commit/7f382097c37c2b3133727729480134abc84bf9c1))
+
 ## [4.0.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-preact@4.0.11...gatsby-plugin-preact@4.0.12) (2020-09-15)
 
 **Note:** Version bump only for package gatsby-plugin-preact