+++ packages/gatsby-remark-responsive-iframe/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-responsive-iframe@2.4.16...gatsby-remark-responsive-iframe@2.4.17) (2020-10-06)
+
+**Note:** Version bump only for package gatsby-remark-responsive-iframe
+
+## [2.4.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-responsive-iframe@2.4.15...gatsby-remark-responsive-iframe@2.4.16) (2020-10-01)
+
+**Note:** Version bump only for package gatsby-remark-responsive-iframe
+
+## [2.4.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-responsive-iframe@2.4.14...gatsby-remark-responsive-iframe@2.4.15) (2020-09-28)
+
+**Note:** Version bump only for package gatsby-remark-responsive-iframe
+
 ## [2.4.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-responsive-iframe@2.4.13...gatsby-remark-responsive-iframe@2.4.14) (2020-09-15)
 
 **Note:** Version bump only for package gatsby-remark-responsive-iframe