+++ packages/gatsby-transformer-hjson/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-hjson@2.4.13...gatsby-transformer-hjson@2.4.14) (2020-10-20)
+
+### Performance Improvements
+
+- implement shouldOnCreateNode for all our plugins/benchmarks ([#27545](https://github.com/gatsbyjs/gatsby/issues/27545)) ([df914d9](https://github.com/gatsbyjs/gatsby/commit/df914d94a7c47c6082b6f165eb44dc6e15e12c7d))
+
+## [2.4.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-hjson@2.4.12...gatsby-transformer-hjson@2.4.13) (2020-10-06)
+
+**Note:** Version bump only for package gatsby-transformer-hjson
+
+## [2.4.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-hjson@2.4.11...gatsby-transformer-hjson@2.4.12) (2020-09-28)
+
+**Note:** Version bump only for package gatsby-transformer-hjson
+
 ## [2.4.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-hjson@2.4.10...gatsby-transformer-hjson@2.4.11) (2020-09-15)
 
 **Note:** Version bump only for package gatsby-transformer-hjson